pyproject.toml
[tool.poetry]
name = "simimibot"
version = "0.1.0"
description = ""
authors = ["Roman Lupashko <lupashcko.rom@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "3.4.1"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
